config: fp16 round arrays default ON (gate passed); bf16 stays off (gate failed)
The pre-agreed val-metric gate, judged on the unregularized configuration
across two batches (distribution-level — the clustering-fn inlining
changed float summation order, so same-seed cross-build pairing is
invalid and stated as such):

- CONTROLS (6 seeds): val AUC .9925-.9990, recall@0.01FPR .9449-.9907,
  active units 6-8 — the raw AU count flips across CONTROL seeds, so the
  'identical AU' criterion is retired in favor of the per-dim variance
  margins (which show marginal-dim wobble, not collapses).
- fp16 (4 seeds): PASS — val >= .9979 on every seed, recalls
  .9619-.9907 all inside the control spread, AU 7-8, zero calibration
  trips. training.round_array_dtype -> 'float16': halves the round
  footprint/gather/page-cache working set; 'float32' restores the
  historical input numerics byte-for-byte (manifest-gated either way).
- bf16 (7 seeds): FAIL — six seeds clean, but seed 13 reproducibly
  degrades (recall .8432 / val .9807 vs the .9449 / .9925 control floor)
  and bf16-seed-13 was the ONLY configuration to trip the
  ECE->calibrator gate, twice, across configs. beta_vae.mixed_precision
  stays False with the verdict pinned in the config comment; revisit
  only with the pathology understood.

+++ src/aetherscan/config.py
@@ -104,7 +104,12 @@ class BetaVAEConfig:
     # variables/optimizer state stay fp32 under keras mixed precision, and the numerically
     # sensitive layers (z_mean/z_log_var/Sampling, decoder sigmoid output → loss math) are
     # pinned fp32 in models/vae.py. Default False = fp32 end-to-end, byte-identical to before
-    # this flag existed (no policy call is made at all).
+    # this flag existed (no policy call is made at all). STAYS OFF after the 7-seed A/B:
+    # 6 seeds clean, but seed 13 reproducibly fails (recall .8432 / val AUC .9807, far below
+    # the 6-seed control envelope, and the only runs to trip the ECE->calibrator gate were
+    # bf16-seed-13, twice, across configs) — a bf16-specific trajectory pathology the gate
+    # exists to catch. Revisit only with that pathology understood (+1.15x step throughput
+    # and ~halved activation VRAM are the upside on the table).
     mixed_precision: bool = False
     # Whether the conv layers' declared L1/L2 penalties are ADDED to the training objective.
     # v1 default False: the declarations were dead code since inception (a custom loop only
@@ -326,11 +331,14 @@ class TrainingConfig:
     # round footprint (~294.5 -> ~147 GB at full scale), the gather volume, and the page-cache
     # working set — the lever that keeps overlapped epochs at page-cache speed once two rounds
     # no longer fit in RAM — at a quantization cost of <= 2^-12 on the [0, 1] log-normalized
-    # inputs. A/B-GATED: changes input numerics; keep "float32" until the val-metric A/B on the
-    # target host shows parity. The gather map upcasts to float32 host-side, so the training
-    # graph and loss math are unchanged either way; labels + lognorm sidecars stay float32.
-    # Recorded in each round's .done manifest and validated on resume/reuse.
-    round_array_dtype: str = "float32"
+    # inputs. DEFAULT FLIPPED to "float16" 2026-07-29 after passing the val-metric A/B gate:
+    # 4 seeds, every score metric inside the 6-seed control spread (val AUC >= .9979, recalls
+    # .9619-.9907 vs control .9449-.9907), AU within the controls' own 6-8 seed variation, no
+    # calibration trips. The gather map upcasts to float32 host-side, so the training graph
+    # and loss math are unchanged either way; labels + lognorm sidecars stay float32.
+    # Recorded in each round's .done manifest and validated on resume/reuse ("float32"
+    # restores the historical behavior byte-for-byte).
+    round_array_dtype: str = "float16"
 
     # Round data pipeline params (disk-backed per-round datasets, see round_data.py)
     round_data_dir: str | None = None  # Defaults to get_training_file_path("round_data") at runtime
